Add instruction metering interface to wasi intefrace Save consumed instruction count on exit

+++ core/iwasm/common/wasm_exec_env.h
@@ -89,7 +89,7 @@ typedef struct WASMExecEnv {
 
 #if WASM_ENABLE_INSTRUCTION_METERING != 0
     /* instructions to execute */
-    int instructions_to_execute;
+    int64 instructions_to_execute;
 #endif
 
 #if WASM_ENABLE_FAST_JIT != 0
